+++ include/ravennakit/ptp/detail/ptp_stats.hpp
@@ -11,7 +11,6 @@
 #pragma once
 
 #include "ravennakit/core/math/sliding_stats.hpp"
-#include "ravennakit/core/util/throttle.hpp"
 
 #include <cstdint>
 
@@ -24,10 +23,9 @@ struct Stats {
     constexpr static int64_t k_clock_step_threshold_seconds = 1;
     constexpr static double k_calibrated_threshold = 0.0018;
 
-    SlidingStats offset {51};
+    SlidingStats offset_from_master {51};
     SlidingStats filtered_offset {51};
     uint32_t ignored_outliers = 0;
-    Throttle<void> trace_adjustments_throttle {std::chrono::seconds(5)};
 };
 
 }  // namespace rav::ptp