+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The ability to configure the PTP domain.
+- Added a callback rav::ptp::Instance::Subscriber::ptp_stats_updated to receive updates with PTP stats.
 
 ### Changed